moved blob to navbar
