+++ src/App.jsx
@@ -13,7 +13,6 @@ import MessageManager from "./components/MessageManager";
 import ProfilePage from "./components/ProfilePage";
 import Navbar from "./components/Navbar";
 import NotificationPanel from "./components/NotificationPanel";
-import ControlBlob from "./components/ControlBlob";
 import "./App.css";
 
 class App extends Component {
@@ -148,11 +147,15 @@ class App extends Component {
             render={() => {
               return (
                 <div>
-                  <Navbar userId={userId} activeTab="home" />
-                  <ControlBlob
+                  <Navbar
                     userId={userId}
                     makeNotification={this.makeNotification}
+                    activeTab="home"
                   />
+                  {/* <ControlBlob
+                    userId={userId}
+                    makeNotification={this.makeNotification}
+                  /> */}
                   <NewsFeed
                     makeNotification={this.makeNotification}
                     userId={userId}
@@ -168,7 +171,11 @@ class App extends Component {
             render={() => {
               return (
                 <div>
-                  <Navbar userId={userId} activeTab="messages" />
+                  <Navbar
+                    userId={userId}
+                    makeNotification={this.makeNotification}
+                    activeTab="messages"
+                  />
                   <MessageManager userId={userId} />
                 </div>
               );
@@ -179,7 +186,11 @@ class App extends Component {
             render={props => {
               return (
                 <div>
-                  <Navbar userId={userId} activeTab="profile" />
+                  <Navbar
+                    userId={userId}
+                    makeNotification={this.makeNotification}
+                    activeTab="profile"
+                  />
                   <ProfilePage
                     userId={userId}
                     handleLogout={this.handleLogout}
@@ -194,7 +205,11 @@ class App extends Component {
             exact
             render={() => (
               <div>
-                <Navbar userId={userId} activeTab="notifications" />
+                <Navbar
+                  userId={userId}
+                  makeNotification={this.makeNotification}
+                  activeTab="notifications"
+                />
                 <NotificationPanel notifications={notifications} />
               </div>
             )}
