--- src/components/ControlBlob/styles.css
@@ -1,161 +1,0 @@
-#navbar {
-  background-color: #91affc;
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  margin: 0;
-}
-
-#navbar ul {
-  display: flex;
-  margin: 0;
-  list-style: none;
-  padding: 0 20px;
-}
-
-#navbar ul li {
-  flex: 1;
-}
-
-#navbar #control-blob {
-  flex: 3;
-}
-
-#navbar ul li a {
-  display: block;
-  padding: 11px 0;
-  margin-bottom: 10px;
-
-  color: white;
-  text-decoration: none;
-  text-transform: uppercase;
-  font-weight: bold;
-  font-size: 20px;
-
-  width: 200px;
-}
-
-#navbar ul li.activeTab a {
-  background-image: url("../../images/Navbar/tab.svg");
-  background-repeat: no-repeat;
-  color: inherit;
-}
-
-#blob {
-  position: absolute;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  top: 0;
-  margin: auto;
-
-  box-sizing: border-box;
-  height: 30vw;
-  width: 30vw;
-  padding: 3vw 7vw;
-  background: linear-gradient(#a4b0fc, white 60%);
-  border-radius: 50%;
-
-  text-align: center;
-  font-size: 2vw;
-  color: white;
-}
-
-#blob-content {
-  width: 100%;
-  height: 100%;
-}
-
-#blob textarea {
-  border: none;
-  overflow: auto;
-  outline: none;
-
-  -webkit-box-shadow: none;
-  -moz-box-shadow: none;
-  box-shadow: none;
-
-  resize: none;
-
-  background-color: rgba(255, 255, 255, 0);
-  width: 100%;
-  height: 80%;
-  font-size: 2vw;
-  color: white;
-}
-
-#circle {
-  position: absolute;
-  left: 0;
-  right: 0;
-  bottom: -26vw;
-  margin: auto;
-
-  height: 39vw;
-  width: 39vw;
-  background-color: rgba(255, 255, 255, 0.8);
-  border-radius: 50%;
-  border: 5px solid #91affc;
-}
-
-#circle .icon {
-  position: absolute;
-  top: 0;
-  right: 0;
-  left: 0;
-  bottom: 0;
-  margin: auto;
-
-  height: 5vw;
-  width: 5vw;
-}
-
-.deg0 {
-  transform: rotate(-90deg) translate(19.5vw) rotate(90deg);
-}
-.deg22_5 {
-  transform: rotate(-67.5deg) translate(19.5vw) rotate(67.5deg);
-}
-.deg45 {
-  transform: rotate(-45deg) translate(19.5vw) rotate(45deg);
-}
-.deg67_5 {
-  transform: rotate(-22.5deg) translate(19.5vw) rotate(22.5deg);
-}
-.deg90 {
-  transform: translate(19.5vw);
-}
-.deg112_5 {
-  transform: rotate(22.5deg) translate(19.5vw) rotate(-22.5deg);
-}
-.deg135 {
-  transform: rotate(45deg) translate(19.5vw) rotate(-45deg);
-}
-.deg157_5 {
-  transform: rotate(67.5deg) translate(19.5vw) rotate(-67.5deg);
-}
-.deg180 {
-  transform: rotate(90deg) translate(19.5vw) rotate(-90deg);
-}
-.deg202_5 {
-  transform: rotate(-67.5deg) translate(-19.5vw) rotate(67.5deg);
-}
-.deg225 {
-  transform: rotate(-45deg) translate(-19.5vw) rotate(45deg);
-}
-.deg247_5 {
-  transform: rotate(-22.5deg) translate(-19.5vw) rotate(22.5deg);
-}
-.deg270 {
-  transform: translate(-19.5vw);
-}
-.deg292_5 {
-  transform: rotate(22.5deg) translate(-19.5vw) rotate(-22.5deg);
-}
-.deg315 {
-  transform: rotate(45deg) translate(-19.5vw) rotate(-45deg);
-}
-.deg337_5 {
-  transform: rotate(67.5deg) translate(-19.5vw) rotate(-67.5deg);
-}
