+++ src/components/Navbar/styles.css
@@ -41,7 +41,7 @@
   background-repeat: no-repeat;
   color: inherit;
 }
-/*
+
 #blob {
   position: absolute;
   left: 0;
@@ -62,6 +62,11 @@
   color: white;
 }
 
+#blob-content {
+  width: 100%;
+  height: 100%;
+}
+
 #blob textarea {
   border: none;
   overflow: auto;
@@ -75,7 +80,7 @@
 
   background-color: rgba(255, 255, 255, 0);
   width: 100%;
-  height: 100%;
+  height: 40%;
   font-size: 2vw;
   color: white;
 }
@@ -84,7 +89,8 @@
   position: absolute;
   left: 0;
   right: 0;
-  bottom: -26vw;
+  top: 0;
+  bottom: 0;
   margin: auto;
 
   height: 39vw;
@@ -153,4 +159,19 @@
 }
 .deg337_5 {
   transform: rotate(67.5deg) translate(-19.5vw) rotate(-67.5deg);
-} */
+}
+
+.photo-upload-button input[type="file"] {
+  display: none;
+}
+
+.photo-upload-button {
+  cursor: pointer;
+
+  padding: 6px 12px;
+  border-radius: 50px;
+  background-color: #a4b0fc;
+
+  color: white;
+  font-size: 14px;
+}
